+++ grpc-service/src/main/java/com/quorum/tessera/client/GrpcClientImpl.java
@@ -1,4 +1,4 @@
-package com.quorum.tessera.node.grpc;
+package com.quorum.tessera.client;
 
 import com.google.protobuf.ByteString;
 import com.quorum.tessera.api.grpc.PartyInfoGrpc;
